requirements.txt
discord.py-self~=2.1.0
environs~=14.6.0
pydantic~=2.12.5
APScheduler~=3.11.2
loguru~=0.7.3